+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 **Fixes**
 
-- Fixed JSON pointers using negative indices. The JSON Pointer specification (RFC 6901) does not allow negative array indexes. We now raise a `JSONPointerIndexError` if a JSON Pointer attempts to resolve an array item with a negative index. See [#115](https://github.com/jg-rp/python-jsonpath/issues/115). For anyone needing JSON Pointers that support negative indexes, set `JSONPointer.min_int_index` to a suitably negative integer, like `JSONPointer.min_int_index = -(2**53) + 1`.
+- Fixed JSON pointers using negative indices. Previously we would resolve negative indices against array-like values, but the JSON Pointer specification (RFC 6901) does not allow negative array indexes. We now raise a `JSONPointerIndexError` if a JSON Pointer attempts to resolve an array item with a negative index. See [#116](https://github.com/jg-rp/python-jsonpath/pull/116). For anyone needing JSON Pointers that support negative indexes, set `JSONPointer.min_int_index` to a suitably negative integer, like `JSONPointer.min_int_index = -(2**53) + 1`.
+- Fixed the JSON Patch `add` operation. Previously we would raise a `JSONPatchError` when adding to an array value with an index equal to the array's length. Now we append to array values when pointing to one past the last index of the array. See [#117](https://github.com/jg-rp/python-jsonpath/issues/117).
 
 ## Version 2.0.0